Change device_event in handle_event() from u16 to usize
When register_event we pass in the value of u32/u64, so we don't
need to convert it to u16 and then pass it to handle_event.

+++ vhost-user-backend/tests/vhost-user-server.rs
@@ -118,7 +118,7 @@ impl VhostUserBackendMut for MockVhostBackend {
 
     fn handle_event(
         &mut self,
-        _device_event: u16,
+        _device_event: usize,
         _evset: EventSet,
         _vrings: &[VringRwLock],
         _thread_id: usize,
